feat: update user entity
