--- src/user/users.ts
@@ -1,10 +1,0 @@
-export interface User {
-  id: string;
-  login: string;
-  password: string;
-  version: number;
-  createdAt: number;
-  updatedAt: number;
-}
-
-export const users: { [id: string]: User } = {};
